From 7af442bd05b6a85195c8b28bd058811af268087e Mon Sep 17 00:00:00 2001 From: googyanas <googyanas@gmail.com> Date: Thu, 18 Dec 2014 22:56:57 +0000 Subject: [PATCH 653/879] sched: Fix inaccurate accounting for real-time task
It is possible that rq->clock_task was not updated in put_prev_task()
in which case we can potentially overcharge a real-time task for time
it did not run. This is because clock_task could be stale and not
represent the exact time real-time task started running.

Fix this by forcing update of rq->clock_task when real-time task
starts running.

Signed-off-by: Srivatsa Vaddagiri <vatsa@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
---
 kernel/sched/rt.c | 10 ++++++++++
 1 file changed, 10 insertions(+)

diff --git a/kernel/sched/rt.c b/kernel/sched/rt.c
index b5358c8..03c3871 100644
--- a/kernel/sched/rt.c
+++ b/kernel/sched/rt.c
@@ -1405,6 +1405,16 @@ static struct task_struct *_pick_next_task_rt(struct rq *rq)
 		rt_rq = group_rt_rq(rt_se);
 	} while (rt_rq);

+	/*
+	 * Force update of rq->clock_task in case we failed to do so in
+	 * put_prev_task. A stale value can cause us to over-charge execution
+	 * time to real-time task, that could trigger throttling unnecessarily
+	 */
+	if (rq->skip_clock_update > 0) {
+		rq->skip_clock_update = 0;
+		update_rq_clock(rq);
+	}
+
 	p = rt_task_of(rt_se);
 	p->se.exec_start = rq->clock_task;

--
2.8.0.rc3